# packages/openapi/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/engineering/energy/ocr-bill": {
      "post": {
        "description": "آپلود قبض PDF/تصویر → OCR → تحلیل pandapower",
        "operationId": "EngineeringController_ocrBill",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "OCR + تحلیل قبض برق (Proxy)",
        "tags": [
          "engineering"
        ]
      }
    },
...
    "/engineering/energy/manual-analyze": {
      "post": {
        "description": "وقتی OCR موفق نشد — داده مستقیم از فرم به Python ارسال می‌شود",
        "operationId": "EngineeringController_manualAnalyze",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "تحلیل قبض با ورود دستی (Proxy)",
        "tags": [
          "engineering"
        ]
      }
    },
    "/engineering/energy/ocr-preview": {
      "post": {
        "operationId": "EngineeringController_ocrPreview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "پیش‌نمایش OCR قبض (Proxy)",
        "tags": [
          "engineering"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xennic Platform API",
    "description": "Xennic Engineering Platform API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "Xennic Team",
      "url": "https://xennic.com",
      "email": "support@xennic.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://xennic.com/terms"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}